Add files via upload
An initial MVP. There's a variety of updates I'm already thinking about, but this works for the meantime. Employees can enter their information and it will be saved to a text file (the name is hardcoded for the moment, but that will likely change at some point). Once everyone is done voting, the store manager can look at the text file to tally up votes (something else that might change later). The program also prevents an employee from voting twice (although there's honestly easy ways around that, but that's something else to work on later).